+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## Commit 0002
+
+- Se agrega Documento asociado al expediente.
+- Se agrega Historial del Expediente.
+- Se agregan Proveedores.
+- Se agregan Establecimientos.
+- Se agrega API de Catálogos.
+- Se actualiza API de Expedientes para registrar historial.
+- Se amplía schema SQL.
+
 ## Commit 0001
 
 - Se crea estructura inicial del repositorio SIGD-ST.